requirements.txt
torchsummary
torchsummaryX
hyperopt
ray
ward-metrics
pyEDFlib
